Move select selector tests to own file
